backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
payplug==1.4.0
pycparser==3.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
gunicorn
psycopg[binary]
dj-database-url
whitenoise